frostsnap_fonts/font_tools/requirements.txt
freetype-py>=2.4.0